Fix: 만료 시각 수정

+++ app/src/main/java/com/easyhz/picly/domain/model/album/AlbumItem.kt
@@ -3,6 +3,6 @@ package com.easyhz.picly.domain.model.album
 data class AlbumItem(
     val regDate: String,
     val tag: String,
-    val expireDate: Long,
+    val expireDate: Double,
     val thumbnailUrl: String,
 )
